refactor: integrate MlUiHelper for project management and streamline dropdown functionality
