+++ app/screens/view/ml_screen.py
@@ -13,19 +13,22 @@ from kivy.metrics import dp
 from kivy.properties import ListProperty
 from kivy.uix.boxlayout import BoxLayout
 from kivy.uix.button import Button
-from kivy.uix.dropdown import DropDown
 from kivy.uix.gridlayout import GridLayout
 from kivy.uix.label import Label
 from kivy.uix.popup import Popup
 from kivy.uix.progressbar import ProgressBar
 from kivy.uix.screenmanager import Screen
-from kivy.uix.textinput import TextInput
 from kivymd.uix.floatlayout import MDFloatLayout
 from kivymd.uix.label import MDLabel
 from kivymd.uix.selectioncontrol import MDCheckbox
 from PIL import Image
 
-from app.screens.utils.additional import BaseScreen, ImageMDButton, MDLabelBtn
+from app.screens.utils.additional import (
+    BaseScreen,
+    ImageMDButton,
+    MDLabelBtn,
+    MlUiHelper,
+)
 from app.screens.utils.custom_logging import get_logger
 from app.screens.utils.db import DB
 from app.screens.utils.ml import KModel, create_config_file, prepare_dataset
@@ -35,7 +38,7 @@ from app.screens.utils.utils import extend_key
 logger = get_logger(__name__)
 
 
-class MLViewScreen(Screen, BaseScreen):
+class MLViewScreen(Screen, BaseScreen, MlUiHelper):
     rgba = ListProperty([1, 1, 0, 0])  # error message popup color
 
     def __init__(self, **kwargs):
@@ -347,7 +350,6 @@ class MLViewScreen(Screen, BaseScreen):
         self.progress_bar.value += 1
         im_path = self.images_to_load.pop(0)
 
-        # 1) The image button
         img = ImageMDButton(
             source=im_path,
             allow_stretch=True,
@@ -807,23 +809,6 @@ class MLViewScreen(Screen, BaseScreen):
                 # btn.allow_hover = True
                 self.ids.model_grid.add_widget(btn)
 
-    def select_model_btn(self, instance):
-        logger.info(f"The model button <{instance.text}> is being pressed")
-        if self.selected_model:
-            if instance.uid == self.selected_model.uid:
-                # custom double touch event
-                self.unselect_model_btn()
-                return
-
-        # reset selection
-        for btn in self.ids.model_grid.children:
-            btn.md_bg_color = (1.0, 1.0, 1.0, 0.0)
-
-        instance.md_bg_color = (1.0, 1.0, 1.0, 0.1)
-        instance.radius = (20, 20, 20, 20)
-        self.selected_model = instance
-        self.update_all_button_states()
-
     def update_all_button_states(self):
         has_selection = bool(self.selected_images)
         can_transfer = (
@@ -926,59 +911,7 @@ class MLViewScreen(Screen, BaseScreen):
 
         self.unselect_all_images()
 
-    def create_project_name_input_popup(self):
-        self.popup = Popup(
-            title="New project creation", size_hint=(None, None), size=(400, 150)
-        )
-        box = BoxLayout(orientation="vertical")
-
-        lbl = Label(text="Please enter new name", size_hint_y=0.3)
-
-        name_input = TextInput(
-            text="",
-            hint_text="Project name",
-            size_hint_y=0.4,
-            multiline=False,
-            font_size=16,
-        )
-        name_input.bind(
-            on_text_validate=lambda x: self.open_project_folder(
-                name_input.text,
-            ),
-        )
-
-        submit_btn = MDLabelBtn(text="Create", size_hint_y=0.3)
-        submit_btn.bind(
-            on_release=lambda x: self.open_project_folder(
-                name_input.text,
-            )
-        )
-        submit_btn.allow_hover = True
-
-        box.add_widget(lbl)
-        box.add_widget(name_input)
-        box.add_widget(submit_btn)
-
-        self.popup.content = box
-        self.popup.open()
-
-    def open_project_folder(self, project_name):
-        logger.debug(f"project_name, {project_name}")
-        if project_name == "":
-            return
-
-        if self.popup is not None:
-            self.popup.dismiss()
-
-        # trigger popup and then call this method again with correct name
-        if project_name == "New project":
-            self.create_project_name_input_popup()
-            return
-
-        self.main_button.text = project_name
-        cur_project_path = os.path.join(self.projects_folder, project_name)
-        os.makedirs(cur_project_path, exist_ok=True)
-
+    def after_project_selection_hook(self, project_name, path):
         folders = [
             "all",
             "configs",
@@ -987,10 +920,8 @@ class MLViewScreen(Screen, BaseScreen):
             "train",
         ]  # TODO: rename train to dataset(s)
         for folder in folders:
-            os.makedirs(os.path.join(cur_project_path, folder), exist_ok=True)
-
-        self.active_project = project_name
-        self.restore_project_params(project_name, cur_project_path)
+            os.makedirs(os.path.join(path, folder), exist_ok=True)
+        logger.debug(f"Subfolders verified/created in {path}")
 
     def restore_project_params(self, project_name, cur_project_path):
         self.loaded_hash = ""
@@ -1004,37 +935,8 @@ class MLViewScreen(Screen, BaseScreen):
         self.show_folder_images(path=os.path.join(cur_project_path, "all"))
 
     def select_project_button(self):
-        projects = []
-        for folder in os.listdir(self.projects_folder):
-            if os.path.isdir(os.path.join(self.projects_folder, folder)):
-                projects.append(folder)
-
-        # If projects folders changed or dropdown was not created
-        if projects != self.projects or self.dropdown is None:
-            if self.dropdown is not None:
-                logger.debug("clear bind")
-                self.main_button.unbind(on_release=self.dropdown.open)
-
-            logger.debug("create bind")
-            self.dropdown = DropDown()
-            for folder in projects:
-                btn = Button(text=f"{folder}", size_hint_y=None, height=44)
-                btn.bind(on_release=lambda b: self.dropdown.select(b.text))
-                self.dropdown.add_widget(btn)
-
-            btn_new = Button(text="New project", size_hint_y=None, height=44)
-            btn_new.bind(on_release=lambda b: self.dropdown.select(b.text))
-            btn_new.background_color = 0.5, 0.9, 0.5, 1
-            self.dropdown.add_widget(btn_new)
-
-            self.main_button.bind(on_release=self.dropdown.open)
-            self.dropdown.bind(
-                on_select=lambda instance, project: self.open_project_folder(project)
-            )
-            self.projects = projects
-
-        else:
-            logger.debug("use bind")
+        projects = self.get_projects()
+        self.setup_project_dropdown(projects)
 
     def on_text_input_class(self, instance, value):
         text = self.ids.class_input.text
